Ensure faf image uses the baked in langtags.json
Specify the include langtag.json explicitly via the LANGTAGSPATH env
variable.  This skips any online checks and is a bit faster to start,
and ensures the langtags.json stays in sync with the sldr data included
in the image. As the image is always rebuilt and redeployed whenever
langtags.json, sldr or fonts/families.json changes the freshness checks
are redundant too.